requirements.txt
# Web API 依賴包
flask>=2.0.0
flask-cors>=3.0.0
torchmetrics>=0.11.0
librosa>=0.10.0

# 已包含在PyTorch環境中的包 (通常不需要額外安裝)
# torch
# torchaudio
# numpy

# 其他可選依賴
werkzeug>=2.0.0
